+++ apps/backend/src/modules/categories/libs/types/category-score-model.type.ts
@@ -1,8 +1,0 @@
-import { type CategoryModel } from "../../category.model.js";
-
-type CategoryScoreModel = {
-	score: number;
-	userId: number;
-} & CategoryModel;
-
-export { type CategoryScoreModel };