+++ apps/mobile/src/libs/constants/constants.ts
@@ -1,1 +1,2 @@
+export { COUNTDOWN_EXPIRED } from "./countdown-expired.constant";
 export { FIRST_ITEM_INDEX, PREVIOUS_INDEX_OFFSET, ZERO_INDEX } from "shared";